+++ mcpscore/rules/__init__.py
@@ -59,11 +59,14 @@ from .readiness import (
     HeaderValidationReadinessRule,
     MetaValidationReadinessRule,
     NoSessionIdReadinessRule,
+    OriginValidationRule,
     RemovedMethodsReadinessRule,
+    ResponseContentTypeRule,
     ResultTypeReadinessRule,
     ServerDiscoverReadinessRule,
     StatelessRequestReadinessRule,
     ToolSchemaDialectReadinessRule,
+    UnknownMethodErrorRule,
     UnsupportedVersionErrorReadinessRule,
 )
 from .registry import RuleRegistry, create_all_rules
@@ -152,6 +155,7 @@ __all__ = (
     "MalformedRequestHandlingRule",
     "MetaValidationReadinessRule",
     "NoSessionIdReadinessRule",
+    "OriginValidationRule",
     "PromptsArgumentNamesPresentRule",
     "PromptsArgumentNamesUniqueRule",
     "PromptsArgumentsDocumentedRule",
@@ -177,6 +181,7 @@ __all__ = (
     "ResourcesTitlesPresentRule",
     "ResourcesUrisUniqueRule",
     "ResourcesUrisValidRule",
+    "ResponseContentTypeRule",
     "ResultTypeReadinessRule",
     "RuleRegistry",
     "RuleResult",
@@ -210,6 +215,7 @@ __all__ = (
     "ToolsOutputSchemaRootObjectRule",
     "ToolsOutputSchemaValidRule",
     "ToolsTitlePresentRule",
+    "UnknownMethodErrorRule",
     "UnsupportedVersionErrorReadinessRule",
     "create_all_rules",
 )